zaakafhandelcomponent-1169 - Document verplaatsen logica naar aparte service
